function_call: avoid parsing calls if no functions are given
In some weird combinations of tools & models (the one that led to us
running into this bug was Qwen3-coder + Roo Code) the tools prompt
models to make function calls in their own alternative syntax.

In the case of Roo Code, this conflicts with the Qwen3 call parser at
the moment.

It seems like a reasonable way to defend against this is to never parse
tool calls if there were no tools provided.