scheduler/output_processor: add DLPM token accounting
Adds accounting for token deficits when DLPM scheduling policies are
used.

For non-speculative decoding, this is done in real time as suggested by
the DLPM paper. In the speculative case, we "commit" to tokens inside of
the speculation workers, where we don't have the scheduler available. In
this case, we update the counters only once the request is finished.