srt/constrained: do not wrap invalid grammar objects
Wrapping invalid grammar objects in the reasoner backend causes the
scheduler's check for invalid grammars to not notice them.

This can lead to requests with invalid grammars (e.g. on regex validity
issues) to be scheduled, eventually causing a crash when the invalid
grammar is used for masking.

This relates to sgl-project/sglang#9187